Switch to "respec-w3c" profile
The previous "respec-w3c-common" has been deprecated. The CG is not affected
by Process 2020, but it is still good practice to use the latest version of
Respec and the old profile will no longer be updated.

Switching to the new profile allows to associate the spec with the group more
easily through the new `group` property, as done in this update.